moved kmeans to new package
